requirements.txt
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.7.0
brotli==1.2.0
bs4==0.0.2
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaitaistruct==0.11
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydeck==0.9.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selenium==4.39.0
selenium-wire==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
websocket-client==1.9.0
wsproto==1.3.2
zstandard==0.25.0
